# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "States of the World API",
    "description":"API for retrieving data about countries, populations, and neighbors.",
    "version": "1.0.0",
    "contact": {
      "name": "Galateanu Petru-Ioan",
      "email": "galateanupetru152@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "paths": {
...
    "/api/countries/search": {
      "get": {
        "summary": "Search countries with filters",
        "description": "Filter countries by language, neighbor, political system, or timezone.",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by language (e.g., 'English')"
          },
          {
            "name": "neighbor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by neighbor (e.g., 'France')"
          },
          {
            "name": "political_system",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by political system (e.g., 'Republic')"
          },
          {
            "name": "timezone",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by timezone (e.g., 'UTC+1')"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matching countries"
          }
        }
      }
    },
    "/api/country/{name}": {
      "get": {
        "summary": "Get Country Details",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed country information"
          },
          "404": {
            "description": "Country not found"
          }
        }
      }
    }
  }
}